pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-kline-vision-data"
version = "4.3.11"
description = "Professional market data integration with clean package architecture"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    {name = "EonLabs", email = "terry@eonlabs.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "attrs>=25.0.0", # data classes and validation
    "pyarrow>=19.0.0,<21.0.0", # Arrow IPC cache (MMAP)
    "polars>=1.30.0,<2.0.0", # internal LazyFrame pipeline
    "pandas>=2.2.0,<3.0.0", # API boundary output (default)
    "numpy>=1.26.0,<3.0.0", # timestamp processing
    "fsspec>=2024.6.0", # Vision API S3 file access
    "requests>=2.32.0", # REST API client + fsspec backend
    "httpx>=0.28.0", # preferred HTTP client
    "tenacity>=9.0.0", # retry with exponential backoff
    "rich>=13.0.0", # formatted logging output
    "pendulum>=3.0.0", # timezone-aware datetime ops
    "platformdirs>=4.0.0", # cross-platform cache paths
    "loguru>=0.7.0", # structured logging
]

[project.optional-dependencies]
dev = [
    "build>=1.2.0", # package distribution builder
    "twine>=6.0.0", # PyPI upload
    "pytest>=8.0.0", # test framework
    "pytest-cov>=6.0.0", # coverage reporting
    "ruff>=0.11.0", # linting and formatting
    "import-linter>=2.0.0", # architectural contract enforcement
    "typer>=0.16.0", # CLI for dev scripts (recmove)
    "GitPython>=3.1.0", # git operations (recmove)
    "rope>=1.0.0", # Python refactoring (recmove)
]
streaming = [
    "websockets>=16.0,<17.0", # async WebSocket client (RFC 6455 compliant)
    "orjson>=3.10", # 3-10x faster JSON parsing than stdlib json
    "python-statemachine>=2.4", # declarative FSM for connection lifecycle
    "stamina>=24.3", # production-grade retry/backoff (wraps tenacity)
]
streaming-test = [
    "hypothesis>=6.115", # property-based testing + RuleBasedStateMachine for FSM
    "time-machine>=2.16", # 400x faster than freezegun, monotonic clock safe
    "pytest-asyncio>=0.24", # async test runner
]
streaming-stress = [
    "memray>=1.14", # native heap profiler (catches Polars/Arrow/Rust allocs)
    "pytest-memray>=1.5", # @pytest.mark.limit_leaks decorator
]
# Command-line scripts defined here
[project.scripts]
recmove = "scripts.dev.refactor_move:app"

[project.urls]
Homepage = "https://github.com/terrylica/crypto-kline-vision-data"
Repository = "https://github.com/terrylica/crypto-kline-vision-data"
Documentation = "https://github.com/terrylica/crypto-kline-vision-data/blob/main/CLAUDE.md"
Changelog = "https://github.com/terrylica/crypto-kline-vision-data/blob/main/CHANGELOG.md"
"Issue Tracker" = "https://github.com/terrylica/crypto-kline-vision-data/issues"

# Using src-layout with proper package structure
[tool.setuptools.packages.find]
where = ["src"]
include = ["ckvd*"]

# Set package directory mapping for src-layout
[tool.setuptools.package-dir]
"" = "src"

# Add package data to ensure all Python files are included
[tool.setuptools.package-data]
"ckvd" = ["**/*.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

# Align with project.requires-python (>=3.13)
# Ruff comment - this cannot be inherited automatically
target-version = "py313"
# Allow autofix for all enabled rules
lint.fixable = ["ALL"]
# Allow unused variables when underscored
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Respect .gitignore
respect-gitignore = true
# Exclude directories from Ruff checks
lint.exclude = ["playground/", "tests/", ".git/", "__pycache__/"]

[tool.basedpyright]
exclude = ["**/__pycache__", "**/.cache", ".venv"]
ignore = ["**/tests/**/*.py", "tests/**/*.py", "ml_feature_set/tests/**/*.py"]
defineConstant = { DEBUG = true }
reportMissingImports = "error"
reportMissingTypeStubs = false
pythonVersion = "3.13"
typeCheckingMode = "basic"

[tool.coverage.run]
branch = true
source = ["src/ckvd"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
show_missing = true

# =============================================================================
# Import Architecture Contracts
# Enforces layering: core → utils (not reverse), provider isolation
# Pre-existing violations documented in ignore_imports (GitHub #16 context)
# =============================================================================

[tool.importlinter]
root_packages = ["ckvd"]
include_external_packages = false

[[tool.importlinter.contracts]]
name = "Layered architecture: core → utils (not reverse)"
type = "layers"
layers = [
    "ckvd.core",
    "ckvd.utils",
]
ignore_imports = [
    "ckvd.utils.for_core.vision_file_utils -> ckvd.core.providers.binance.rest_data_client",
    "ckvd.utils.for_core.ckvd_cache_utils -> ckvd.core.providers.binance.vision_path_mapper",
]

[[tool.importlinter.contracts]]
name = "Provider isolation: Binance and OKX are independent"
type = "independence"
modules = [
    "ckvd.core.providers.binance",
    "ckvd.core.providers.okx",
]
# DataClientInterface is a shared ABC that should live in core/providers/
# but currently resides in binance/. Moving it is a separate refactoring task.
ignore_imports = [
    "ckvd.core.providers.okx.okx_rest_client -> ckvd.core.providers.binance.data_client_interface",
]

[[tool.importlinter.contracts]]
name = "No deleted layers: for_demo must stay deleted"
type = "forbidden"
source_modules = [
    "ckvd",
]
forbidden_modules = [
    "ckvd.utils.for_demo",
    "ckvd.utils_for_debug",
]